requirements.txt
scikit-learn
biopython==1.79
numpy==1.14.5
#tensorflow-gpu==1.10.0
scipy==1.2.0
theano==1.0.5
evcouplings==0.1.1
